Default background filename standardised as be.dat
rsl files only moved around at the end of a if they exist
(they don't for serial ones, which generated an error)

bluevista:afwa_swa gen_be tested

No scientific change

Reviewed by Mike Duda

M    gen_be/gen_be_ep1.f90
M    gen_be/gen_be_diags.f90
M    gen_be/gen_be_diags_read.f90
M    scripts/gen_be/gen_be_ep1.csh
M    scripts/gen_be/gen_be_etkf.csh
M    scripts/da_run_wpb.ksh
M    scripts/da_run_suite.ksh
M    scripts/da_run_wrfvar.ksh
M    scripts/da_perturb_wrf_bc.ksh
M    scripts/da_run_etkf.ksh


git-svn-id: https://svn-wrf-var.cgd.ucar.edu/trunk@2465 70815b0f-b814-0410-81f7-dfa10a885da6

+++ scripts/da_run_wrfvar.ksh
@@ -64,7 +64,7 @@ if $CYCLING; then
 fi
 
 export DA_ANALYSIS=${DA_ANALYSIS:-analysis}
-export DA_BACK_ERRORS=${DA_BACK_ERRORS:-$BE_DIR/gen_be.NMC.dat} # wrfvar background errors.
+export DA_BACK_ERRORS=${DA_BACK_ERRORS:-$BE_DIR/be.dat} # wrfvar background errors.
 
 export RTTOV=${RTTOV:-$HOME/rttov/rttov87}                            # RTTOV
 export DA_RTTOV_COEFFS=${DA_RTTOV_COEFFS:-$RTTOV/rtcoef_rttov7}
@@ -609,12 +609,14 @@ echo "WINDOW_END            $WINDOW_END"
       mkdir -p $RUN_DIR/rsl
       cd $RUN_DIR/rsl
       for FILE in $WORK_DIR/rsl*; do
-         FILE1=`basename $FILE`
-         echo "<HTML><HEAD><TITLE>$FILE1</TITLE></HEAD>" > $FILE1.html
-         echo "<H1>$FILE1</H1><PRE>" >> $FILE1.html
-         cat $FILE >> $FILE1.html
-         echo "</PRE></BODY></HTML>" >> $FILE1.html
-         rm $FILE
+         if test -f $FILE; then
+            FILE1=`basename $FILE`
+            echo "<HTML><HEAD><TITLE>$FILE1</TITLE></HEAD>" > $FILE1.html
+            echo "<H1>$FILE1</H1><PRE>" >> $FILE1.html
+            cat $FILE >> $FILE1.html
+            echo "</PRE></BODY></HTML>" >> $FILE1.html
+            rm $FILE
+         fi
       done
       cd $RUN_DIR
